refactored app boot into multiple files, added proper backend process killing on app quit
